+++ tests/fixtures.py
@@ -1,0 +1,42 @@
+import pathlib
+import random
+import string
+import tempfile
+import discord
+import pytest
+import os
+from word_debt_bot import core
+
+from word_debt_bot.game.core import WordDebtGame
+from word_debt_bot.game.player import WordDebtPlayer
+
+
+@pytest.fixture
+def filepath() -> pathlib.Path:
+    file_path = tempfile.mkstemp()
+    yield pathlib.Path(file_path[1])
+    os.remove(file_path[1])
+
+
+@pytest.fixture
+def game(filepath: pathlib.Path) -> WordDebtGame:
+    return WordDebtGame(filepath)
+
+
+@pytest.fixture
+def player() -> WordDebtPlayer:
+    user_id = str(random.randint(10**7, 10**8 - 1))
+    user_name = "".join(random.choice(string.ascii_lowercase) for _ in range(8))
+    return WordDebtPlayer(user_id, user_name)
+
+
+@pytest.fixture
+def bot(game: core.WordDebtGame, filepath: pathlib.Path) -> core.WordDebtBot:
+    intents = discord.Intents.default()
+    intents.message_content = True
+
+    bot = core.WordDebtBot(command_prefix='.', intents=intents)
+    bot.game = game
+    bot.journal_path = filepath
+
+    return bot
